+++ src/meeple/command/search.py
@@ -1,7 +1,15 @@
+import sys
+
 import click
 
 from meeple.util.api_util import search_bgg
-from meeple.util.output_util import ItemHeader, fmt_headers, fmt_year, print_table
+from meeple.util.output_util import (
+    ItemHeader,
+    fmt_headers,
+    fmt_year,
+    print_table,
+    print_warning,
+)
 
 
 @click.command()
@@ -13,15 +21,24 @@ def search(query: str) -> None:
     - QUERY is the text to be searched for on BoardGameGeek. If searching multiple words, surround with quotes.
     """
     # search BoardGameGeek with user provided query
-    api_result = search_bgg(query)
-    api_result.sort(key=lambda x: x.id)
+    result_items = search_bgg(query)
+
+    # check that data exists after applied filters
+    if not result_items:
+        sys.exit(
+            print_warning(
+                f"No items found on BoardGameGeek matching search term [i blue]{query}[i blue]."
+            )
+        )
+
+    result_items.sort(key=lambda x: x.id)
 
     # prepare table data
     headers = [ItemHeader.ID, ItemHeader.NAME, ItemHeader.YEAR]
     headers = fmt_headers(headers, None, None)
 
     rows = []
-    for item in api_result:
+    for item in result_items:
         cols = []
         cols.extend(
             [